ExerciseDialog: implemented mapping of sport types and subtypes when importing FIT files

+++ st-exerciseviewer/src/main/kotlin/de/saring/exerciseviewer/data/SportTypeFit.kt
@@ -11,6 +11,6 @@ package de.saring.exerciseviewer.data
  * @author Stefan Saring
  */
 data class SportTypeFit(
-    var sportTypeId: Short,
-    var sportSubTypeId: Short? = null,
+    var sportTypeId: Int,
+    var sportSubTypeId: Int? = null,
 )
